new: recreated the main module "start.py" improved: "start.py" can import and work with "api_processor.py"
