--- settings.json
@@ -1,1 +1,0 @@
-{"repo": "freeCodeCamp/freeCodeCamp"}
